requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
pydantic-settings>=2.0
sentence-transformers>=2.6
spacy>=3.7
httpx>=0.27
python-dotenv>=1.0
langdetect>=1.0
prometheus-eval
litellm
pytest>=8.0
pytest-asyncio>=0.23
ruff>=0.4
